GHI #32 Remove some static casts from BtTypesIds

+++ test/kind/bt/types/ids.cpp
@@ -183,13 +183,13 @@ TEST_F(BtTypesIds, get_ids_gives_correct_ids_for_all_kind_combinations)
 {
     // calculate all kind combinations
     std::vector<int> all_kind_combos;
-    all_kind_combos.resize(static_cast<std::size_t>(1) << kinds.size());
+    all_kind_combos.resize(1u << kinds.size());
     for (std::size_t i = 0; i < all_kind_combos.size(); ++i)
     {
         int kind_combo = 0;
         for (std::size_t j = 0; j < kinds.size(); ++j)
         {
-            if (i & (static_cast<std::size_t>(1) << j))
+            if (i & (1u << j))
             {
                 kind_combo |= kinds[j];
             }
@@ -197,7 +197,7 @@ TEST_F(BtTypesIds, get_ids_gives_correct_ids_for_all_kind_combinations)
         all_kind_combos[i] = kind_combo;
     }
 
-    // skip empty set (in case there is no valid kind with value 0)
+    // skip empty set
     all_kind_combos.erase(all_kind_combos.begin());
